Move matchPropertyByURL into utils folder.
