+++ assets/js/modules/analytics/datastore/accounts.js
@@ -36,6 +36,7 @@ import { actions as errorStoreActions } from '../../../googlesitekit/data/create
 import { actions as tagActions } from './tags';
 import { actions as propertyActions } from './properties';
 import { CORE_SITE } from '../../../googlesitekit/datastore/site/constants';
+import { matchPropertyByURL } from '../util/property';
 const { createRegistrySelector } = Data;
 const { receiveError, clearError } = errorStoreActions;
 
@@ -145,7 +146,7 @@ const baseActions = {
 			const urls = registry.select( CORE_SITE ).getSiteURLPermutations();
 			const uaProperties = registry.select( STORE_NAME ).getProperties( accountID );
 
-			let uaProperty = yield propertyActions.matchPropertyByURL( uaProperties, urls );
+			let uaProperty = matchPropertyByURL( uaProperties, urls );
 			if ( ! uaProperty ) {
 				uaProperty = {
 					id: PROPERTY_CREATE,
