+++ assets/js/modules/analytics/datastore/properties.js
@@ -32,7 +32,6 @@ import { STORE_NAME, PROPERTY_CREATE, PROFILE_CREATE } from './constants';
 import { createFetchStore } from '../../../googlesitekit/data/create-fetch-store';
 import { actions as errorStoreActions } from '../../../googlesitekit/data/create-error-store';
 import { MODULES_ANALYTICS_4 } from '../../analytics-4/datastore/constants';
-import { normalizeURL } from '../../../util';
 
 // Get access to error store action creators.
 // If the parent store doesn't include the error store,
@@ -247,34 +246,6 @@ const baseActions = {
 			payload: { primaryPropertyType },
 			type: SET_PRIMARY_PROPERTY_TYPE,
 		};
-	},
-
-	/**
-	 * Matches a property by URL.
-	 *
-	 * @since n.e.x.t
-	 *
-	 * @param {Array.<Object>}        properties Array of properties.
-	 * @param {Array.<string>|string} url        A list of URLs or a single URL to match properties.
-	 * @return {Object|null} A property object if found. NULL if nothing is found.
-	 */
-	*matchPropertyByURL( properties, url ) {
-		const urls = ( Array.isArray( url ) ? url : [ url ] ).map( normalizeURL );
-
-		for ( const property of properties ) {
-			const websiteURL = property.websiteUrl; // eslint-disable-line sitekit/acronym-case
-			if ( ! websiteURL ) {
-				continue;
-			}
-
-			for ( const singleURL of urls ) {
-				if ( singleURL === normalizeURL( websiteURL ) ) {
-					return property;
-				}
-			}
-		}
-
-		return null;
 	},
 };
