+++ assets/js/modules/analytics/util/property.js
@@ -1,0 +1,50 @@
+/**
+ * Property helpers.
+ *
+ * Site Kit by Google, Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Internal dependencies
+ */
+import { normalizeURL } from '../../../util/urls';
+
+/**
+ * Matches a property by URL.
+ *
+ * @since n.e.x.t
+ *
+ * @param {Array.<Object>}        properties Array of properties.
+ * @param {Array.<string>|string} url        A list of URLs or a single URL to match properties.
+ * @return {Object|null} A property object if found. NULL if nothing is found.
+ */
+export function matchPropertyByURL( properties, url ) {
+	const urls = ( Array.isArray( url ) ? url : [ url ] ).map( normalizeURL );
+
+	for ( const property of properties ) {
+		const websiteURL = property.websiteUrl; // eslint-disable-line sitekit/acronym-case
+		if ( ! websiteURL ) {
+			continue;
+		}
+
+		for ( const singleURL of urls ) {
+			if ( singleURL === normalizeURL( websiteURL ) ) {
+				return property;
+			}
+		}
+	}
+
+	return null;
+}
